requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
groq
torch
transformers
